chore: move DEFAULT_FAVOURITE_TOKENS to separate file
